refactor: modal 컴포넌트 분리
